Add Day 3: Conditionals, logic operators, and nested statements
Revisited if, else, and elif statements with comparison operators, which I already knew thanks to Dr. Chuck, but it was great to refresh and actually apply them in longer examples. The % module was central to the day and used in creative ways. I also learned more about logical operators and how 'not' really works. I used to think it just meant “if not that,” but I now get how it reverses the truth value of the condition. The pizza and rollercoaster examples were fun and helped me see how nested conditionals play out. Treasure Island was a cool challenge, simple in logic but took longer than expected due to writing the story part.