Add Day 5: For loops, range(), and password generator project
Already knew how for loops worked, but the range() function was new to me or something I had completely forgotten. Learning how to combine it with for loops opened a lot of possibilities, and it felt familiar thanks to past exercises like manually finding a max value with loops and conditions. The password generator project was the real challenge. I kept trying to use only one loop with logic to check all conditions, but finally settled on using three separate for loops based on user input. I also used new tools like "".join() to concatenate the password list, and random.shuffle() to randomize the order. I’m really starting to appreciate how useful the random module is, especially with choice().