Update Day7.py
Finally finished the Day 7 of the hangman game and it was really fun to do. The while, for and if loops/statement's were used a lot in this one but I'm proud of taking the challenge.
Also I did improvement in a few things that were not covered by Angela making sure there where only one letter as input and it was an alphabet character.