Create Day9.py
This day was to create a biding auction and it was quite simple because I already knew how to use dictionaries and also because I was not afraid of googling functions to make what I envision.
I still have to polish this code though for the float number for the biding but I will tackle that problem later.