Create Day11.py
Started the game of Blackjack for my capstone project for day 11.
It has come up easy I just have to think on how to make the loop for keep on making a choice for the random card for each response of the user when they want to get more cards on their deck.